[Submit] Break out into action and command
